backend/requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP Processing
spacy>=3.5.0
nltk>=3.8.0

# PDF and Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Bias Detection
fairlearn>=0.8.0
aif360>=0.5.0
